Move NodesPackage back into utils
