--- ironflow/main/nodes_package.py
@@ -1,20 +1,0 @@
-from os.path import basename, dirname, splitext, normpath, join
-
-
-class NodesPackage:
-    """
-    A small container to store meta data about imported node packages.
-    """
-
-    def __init__(self, directory: str):
-
-        self.name = basename(normpath(directory))
-        self.directory = directory
-
-        self.file_path = normpath(join(directory, 'nodes.py'))
-
-    def config_data(self):
-        return {
-            'name': self.name,
-            'dir': self.directory,
-        }
